refactor: refactor chatarea
